general: Move Item class to its own files
Keeps implementations within another file to reduce the amount of files
that need to be rebuilt when a change occurs.

Also uncovers some locations in code that were depending on includes
inside other headers instead of explicitly listing its dependencies.
