+++ source/item_list.hpp
@@ -1,135 +1,6 @@
 #pragma once
-#include "logic.hpp"
-#include "random.hpp"
-#include <string>
-#include <functional>
-#include <variant>
 
-#include "../code/src/item_override.h"
-
-enum ItemType {
-    ITEMTYPE_ITEM,
-    ITEMTYPE_MAP,
-    ITEMTYPE_COMPASS,
-    ITEMTYPE_BOSSKEY,
-    ITEMTYPE_SMALLKEY,
-    ITEMTYPE_TOKEN,
-    ITEMTYPE_FORTRESS_SMALLKEY,
-    ITEMTYPE_EVENT,
-    ITEMTYPE_DROP,
-    ITEMTYPE_REFILL,
-    ITEMTYPE_SONG,
-    ITEMTYPE_SHOP,
-    ITEMTYPE_DUNGEONREWARD
-};
-
-class Item {
-public:
-    using EffectFn = void (*)();
-
-    Item(std::string name_, ItemType type_, int getItemId_, bool advancement_, bool* logicVar_, u16 price_ = 0)
-        : name(std::move(name_)),
-          type(type_),
-          getItemId(getItemId_),
-          advancement(advancement_),
-          logicVar(logicVar_),
-          price(price_) {}
-
-    Item(std::string name_, ItemType type_, int getItemId_, bool advancement_, u8* logicVar_, u16 price_ = 0)
-        : name(std::move(name_)),
-          type(type_),
-          getItemId(getItemId_),
-          advancement(advancement_),
-          logicVar(logicVar_),
-          price(price_) {}
-
-    void ApplyEffect() {
-        if (std::holds_alternative<bool*>(logicVar)) {
-          *std::get<bool*>(logicVar) = true;
-        } else {
-          *std::get<u8*>(logicVar) += 1;
-        }
-        Logic::UpdateHelpers();
-    }
-
-    void UndoEffect() {
-        if (std::holds_alternative<bool*>(logicVar)) {
-          *std::get<bool*>(logicVar) = false;
-        } else {
-          *std::get<u8*>(logicVar) -= 1;
-        }
-        Logic::UpdateHelpers();
-    }
-
-    ItemOverride_Value Value() const {
-        //just a random list of progression items
-        const std::array<u16, 10> items = {
-          GI_BOOMERANG,
-          GI_HAMMER,
-          GI_ARROW_LIGHT,
-          0x80,
-          0x81,
-          0x82,
-          0x83,
-          0x84,
-          0x85,
-          0x8A,
-        };
-
-        ItemOverride_Value val;
-        val.all = 0;
-        val.itemId = getItemId;
-        if (getItemId == GI_ICE_TRAP) {
-            val.looksLikeItemId = RandomElement(items);
-        }
-        return val;
-    }
-
-    std::string_view GetName() const {
-        return name;
-    }
-
-    bool IsAdvancement() const {
-        return advancement;
-    }
-
-    int GetItemID() const {
-        return getItemId;
-    }
-
-    ItemType GetItemType() const {
-        return type;
-    }
-
-    u16 GetPrice() const {
-        return price;
-    }
-
-    void SetAsPlaythrough() {
-        playthrough = true;
-    }
-
-    bool IsPlaythrough() const {
-        return playthrough;
-    }
-
-    bool operator== (const Item& right) const {
-        return type == right.GetItemType() && getItemId == right.GetItemID();
-    }
-
-    bool operator!= (const Item& right) const {
-        return type != right.GetItemType() || getItemId != right.GetItemID();
-    }
-
-private:
-    std::string name;
-    ItemType type;
-    int  getItemId;
-    bool advancement;
-    std::variant<bool*, u8*> logicVar;
-    u16  price;
-    bool playthrough = false;
-};
+#include "item.hpp"
 
 extern Item NoItem;
 extern Item I_Triforce;
